Add preprocessor conditional group TIMEKERNELS
This commit adds #IFDEF preprocessor statements in SWE_WavePropogationBlockCuda.cu in order to toggle timing code for computeNetUpdatesKernel, updateUnknownsKernel, and updateUnknownsCUBLAS.  This code makes these kernels each execute 20 times, then prints the result of the timing to standard out.  sys/time.h is used for timing, which should give high-resolution time on Linux.  Note that the time returned is the total execution time of all 20 iterations of the kernels in microseconds.  cudaThreadSynchronize() is used liberally to ensure kernel execution has completed before and after time is ended and begun.  This approach is known to slightly over-estimate execution time, as kernel initialization, return, and synchronization overheads are included.  I think this is okay for evaluating the real clock time of the kernel.