Optimize updateUnknownsCUBLAS to utilize multiple GPU streams.
GPU with compute capability 2.0 and higher are capable of concurrently executing as many as 16 kernels simultaneously.  This optimization targets intermediate-sized problems (length of a row maybe less than 10,000).  Numerical tests show that this optimization didn't affect the numerical behavior of the kernel (compared against unmodified code as of the previous commit).